Rename BuildTask.add_makefile_commands to _add_makefile_commands.
This method is not particularly useful outside of the
BuildTask.makefile_text implementation, even for testing purposes (the
way to extract information from sourcery.makefile.Makefile objects
being to generate makefile text from them).  Thus, rename it to
indicate it's an internal piece of the implementation of another
method, not intended to be called or tested separately.