Add check-style script to run various checkers on the code.
This serves to document the current set of checkers and associated
configuration used in testing Sourcery Builder patches; these checks
are clean with current git versions of pyflakes, pycodestyle and
pylint.  It may be appropriate to integrate these checks better with
the main self-tests in future, either as part of "sourcery-builder
self-test" or as a separate command such as "sourcery-builder
self-test-style".