Add checks on types of release config variables.
A new class ConfigVarType (and subclasses thereof) is added to
describe possible types of values for release config variables, and
how to convert the values for storage.  The base class ConfigVarType
covers cases where an isinstance check suffices, while cases with more
complicated checks and conversions are covered by subclasses, with
ConfigVarTypeList being provided for lists of elements of a given type
(converted to tuples for storage, to avoid subsequent accidental
changes to values) and ConfigVarTypeDict for mappings with given types
of keys and values, as well as ConfigVarTypeStrEnum for strings with
some enumerated set of valid values.  Tests are added for these
classes (but not yet for their users).

ConfigVar and ConfigVarGroup objects start storing their names for use
in diagnostics of invalid types.