Add "sourcery-builder release".
"sourcery-builder release" is like "sourcery-builder devel", but for
release builds.  The intent is that it will be less flexible about
configurability from the command line (so that builds are fully
reproducible from only the set of sources involved), will have extra
sanity checks that the correct sources are being built, and will have
different defaults for some aspects of the build (for example, it will
build source packages, and strip binaries before packaging).  For now,
it enables re-exec of sourcery-builder to ensure a controlled
environment, but most of the additional checks and features are not
yet implemented.