Add support for host-independent build tasks provided by components.
These are for parts of the build that are logically independent of any
particular host, such as building source packages.  As well as general
host-independent tasks there are special cases of init and fini tasks
(formally not needed, but provide dependencies automatically for
convenience in certain use cases; for example, running testsuites as a
fini task once the whole toolchain build is done).