Add support for building source packages.
Source packages are built by "sourcery-builder release", but not by
"sourcery-builder devel".  There is a source package for the open
source components, and a backup package, which is equivalent but for
closed source components.  Component sources are copied (and
permissions cleaned up for reproducibility of the resulting packages)
in init tasks, and per-component tarballs created in host-independent
tasks; host-independent tasks then package those tarballs into a
higher-level source or backup package tarball.