Add release_configs and sourcery_builder components.
These components are for when Sourcery Builder and the release configs
are themselves components checked out with "sourcery-builder
checkout", as required for them to be included in the packaged
sources.  (The actual sources used for the release_configs component
are inherently site-specific, but the existence and some of the logic
around such a component are generic.)

Subsequent changes will add further logic relating to these
components, including arranging for the checkout command (given an
appropriate site-specific wrapper) to be able to bootstrap a checkout
of the correct branches of these components and use those checkouts to
check out all other components.