Support <branch>:<config> release config names.
The intent of such names is that, on checkout, the indicated branch
will (given a suitable subclass of ReleaseConfigPathLoader in use) be
used to checkout the scripts and configs which are then used for the
main checkout of the toolchain.  This bootstrap step is, however, not
implemented in this change, which adds the other part of such support,
as appropriate for commands other than "checkout": using the branch
name to locate a config in an appropriate checkout of release_configs
(which must already exist) in srcdir.  The bootstrap_components_* and
script_full release configs variables are also set automatically,
again given a subclass of ReleaseConfigPathLoader (which needs to set
bootstrap_components to the desired list of components, as well as
defining the branch_to_vc method).