Rework handling of toolchain builds for later hosts.
Target libraries should only be built once, and building host binaries
for a host other than the build system should not depend on other such
host binaries having been built (for example, building GCC can depend
on having binutils or GCC built for the build system, but not on
having binutils built for that later host because binaries built for
that later host can't be run during the build process).  Thus, there
is no need for the toolchain-* install trees for hosts other than the
build system; this patch eliminates them.  PATH settings for the GCC
build are made to use tools for the build system not the host (putting
tools for another host in the PATH does not make sense), and the
build-sysroot setting is made to point to the relevant toolchain trees
for the build system (pointing to such trees for the host there is
harmless if they exist, but unnecessary).

This does not by itself suffice to allow toolchain builds for multiple
hosts to work; further changes are needed for that.