Fix GCC build for hosts other than the build system.
Use all-host and install-host targets to avoid any attempt to rebuild
target libraries.  Use toolchain-2 rather than toolchain-2-before in
dependencies for such hosts because even when not building target
libraries, the compiler for the target is required (e.g. to run with
-dumpspecs), and it must be the final one, not that from toolchain-1,
to get the correct specs.

This does not include any code to make the packages for such hosts
include the target libraries built with GCC (and a few other files
built and installed with those libraries, such as translated messages
and info manuals for those libraries); that will be handled
separately.