Reuse most library files built with GCC for build system for other hosts.
This makes packages for other hosts include most of the GCC-provided
library-related files from the build system (libraries, headers,
libstdc++ GDB Python pretty-printers, info documentation for
libraries, locale .mo files for libstdc++).  It does not handle
headers from libsubdir/include (where some such files are installed
from the gcc/ directory, so present on both hosts, but others are
installed from library directories, so present only on the build
system - so meaning a version of FSTreeUnion that allows duplicate
files where contents are identical will be most appropriate for
handling that directory).  Because libsubdir/include is not yet
handled, README is not updated to reflect support for multiple hosts
either.