Add MapFSTree.union support for allowing duplicate files.
This is for cases where it is natural to take the union of overlapping
sets of files, but files appearing in both places should still be
checked to be identical.  For example, multiple glibc multilibs should
share a single set of installed headers, but glibc generates
gnu/lib-names-<abi>.h and gnu/stubs-<abi>.h headers in multi-ABI
cases; by taking the union and allowing duplicates, all such headers
can be collected in a single place without Sourcery Builder needing to
know the full set of such headers (which may depend on the glibc
version) or the ABI names used by glibc.