Validate source_type, version, vc settings for components.
source_type must be set, not left as the default None, for all
components.  version and vc must be set, not left as the default None,
for components whose source_type is not 'none'.

This check is of particular use in preparation for adding support for
component dependencies, since components being added implicitly as a
result of such dependencies makes it more likely that such settings
for them might be missing in release configs, and so more desirable
that this problem is detected with a meaningful diagnostic given as
soon as possible.