Add support for component dependencies.
The expected context in which component dependencies are useful is
where there is a common include file shared by multiple release
configs that use different subsets of the components from it; those
configs can add just those components that are user-visible parts of
the release and others such as host libraries are then added
automatically through dependencies.

This patch only adds the general support, not uses of it for existing
components; the uses of the component dependency support will be added
separately.