Support wrappers providing their own Component and Command methods.
Site-specific wrappers for Sourcery Builder may have their own
Component or Command methods that are not used by the generic code.
In that case, they need to support site-specific components and
commands, and non-site-specific ones that have a site-specific
definition (inheriting appropriately) to define those site-specific
methods, and non-site-specific ones that do not need special
definitions of those site-specific methods but do need to inherit the
default definitions of those methods.

Thus, expect each package used to have its own component and command
modules (with Component and Command classes) and define a standard
inheritance ordering, with classess constructed dynamically when
needed to ensure the class actually used for a component or command
inherits from all the right classes in the right order.