Add sourcery.package.resolve_symlinks.
This is part of preparation for replacing symlinks by hard links in
binary packages for Windows host.  Symlink resolution is implemented
here in Python, rather than simply following symlinks when copying the
package-input tree, to ensure various kinds of problems with symlinks
are reliably detected.  In particular, symlinks to absolute paths can
be appropriate in intermediate steps in the build (for sysroots that
are not themselves part of the build, for example), and doing such
explicit checks for them avoids the files pointed to being
accidentally copied and packaged.