Support multilib builds of glibc.
This arranges for glibc to be built and installed in the right place
for each glibc multilib, with a single unified set of headers for each
relevant sysroot headers suffix.  It does not deal with the case of
more than one multilib sharing a sysroot, which requires additional
logic to allow for shared files that are identical in such
configurations while moving shared files that are different (e.g.,
executables in sbin/ usr/bin/ usr/sbin/) to directories that do not
conflict between multilibs (such as usr/lib/bin/ and usr/lib64/bin/).