Make MapFSTree / FSTree remove / extract check for passing single string.
Detects the mistake of passing 'some-path' instead of ['some-path'] or
('some-path',), with a more specific error than might otherwise be
given if the path contains '.' or '/' and so is invalid even when
considered as a list of single-character paths (in the absence of
those characters, no error would otherwise have been given for that
mistake).  Similar to such detection in other contexts, e.g. for lists
of command-line options.