Add MapFSTree.extract_one.
This allows creating a MapFSTree to refer directly to a subdirectory
(or file or symlink) in another MapFSTree.  Although formally
redundant with naming appropriate subdirectories when creating a
MapFSTreeCopy, this can nevertheless be convenient.  Intended use
cases include extracting executables from directories such as usr/bin
in sysroots that are shared between all multilibs in a sysroot and
moving them to directories such as usr/lib/bin that can have
per-multilib names such as usr/lib64/bin, to avoid conflicts between
such executables for different multilibs.