Add Multilib.move_sysroot_executables.
This is for the case where a sysroot is shared between multilibs, and
so different multilibs have different library directories such as
usr/lib and usr/lib64, but executables go in the same directory such
as usr/bin for all multilibs.  To avoid conflicts between files for
different multilibs, those executables are moved to per-multilib
directories such as usr/lib/bin (this is an arrangement for packaging,
with users of the sysroot expected to copy the preferred version of a
binary back into directories such as usr/bin).  For user convenience,
copies of the files are left in their original directories if there is
only one multilib in the sysroot.