Support more than one glibc multilib sharing a sysroot.
This is the case of e.g. 32-bit and 64-bit multilibs using lib and
lib64 directories in the same sysroot.  In this case, certain
directories need to be allowed to have identical contents for each
such multilib, while executables need to be moved from directories
such as usr/bin to those such as usr/lib/bin (with copies being left
in the former directory in the case where only one multilib uses a
sysroot).