Add support for building glibc locales.
Locales are built via building host glibc to provide a suitable
localedef binary, and need to be built once per endianness.  Rather
than require release configs to specify endianness of each locale,
this is computed by testing the compiler.  That in turn means the set
of locales to build can only be determined after the compiler has been
built, so building locales in parallel involves generating a makefile
(after compilers have been built) which is then used to build locales.
Similarly, the particular endianness of locale data to use for a given
multilib cannot be determined at the time FSTree objects for install
trees are created, so a glibc-local subclass _FSTreeLocale is used to
allow that choice to be delayed until the relevant install trees need
to be exported.