Support tool_opts as Multilib argument.
In most cases, this is not needed, because either tools do not need
multilib-specific argument, or, if they do (e.g. as and ld), they are
called via the compiler and so specifying ccopts suffices.  However,
windres (as used for Windows target) is a tool that both needs a
multilib-specific argument (such as -F pe-i386) in certain cases, and
is not called via the compiler, so providing a use case for tool_opts
as a Multilib argument (passed through to BuildCfg).