Disable pylint import-outside-toplevel for one self-test.
This test is deliberately using local imports for assertions comparing
results of other operations with names in the imported modules,
instead of the normal case of importing a module to use its
functionality, so the local imports are appropriate.

This gets the results of check-style with current pylint clean again.